로그인 체크 추가

+++ src/components/ReadBoardComponent.jsx
@@ -3,6 +3,7 @@ import BoardService from '../services/BoardService';
 import { useParams } from 'react-router-dom';
 import { useNavigate } from "react-router-dom";
 import dayjs from 'dayjs';
+import * as common from "../utils/common_function";
 
 function ReadBoardComponent() {
     const [board, setBoard] = useState({});
@@ -17,7 +18,7 @@ function ReadBoardComponent() {
                 const data = JSON.parse(response.data);
                 setBoard(data);
             } else {
-
+                alert(response.resultMessage);
             }
         });
     }, []);
@@ -60,6 +61,13 @@ function ReadBoardComponent() {
      * 삭제처리
      */
     const deleteBoard = () => {
+        if (!common.checkLogin()) {
+            alert("장시간 입력이 없어 정보를 가져올 수 없습니다. 다시 진행해주세요.");
+            sessionStorage.setItem("IS_LOGIN", "N");
+            navigate("/");
+            return false;
+        }
+
         if (window.confirm("정말 삭제하시겠습니까?")) {
             BoardService.deleteBoard(no).then(res => {
                 let response = res.data;
@@ -67,7 +75,11 @@ function ReadBoardComponent() {
                     alert("삭제되었습니다.");
                     navigate("/board");
                 } else {
-
+                    alert(response.resultMessage);
+                    if (response.apiResultCode === "0099") {
+                        sessionStorage.setItem("IS_LOGIN", "N");
+                        navigate("/");
+                    }
                 }
             });
         }
